Moved HDF5KeyTranslator back to plugins directory, for now
